+++ doc/changelog.md
@@ -13,6 +13,7 @@ Jump to:
 
 Description
 
+- Enable hostname selection for dragon tasks
 - Remove pydantic dependency from MLI code
 - Update MLI environment variables using new naming convention
 - Reduce a copy by using torch.from_numpy instead of torch.tensor